+++ anymail/urls.py
@@ -4,6 +4,7 @@ from .webhooks.amazon_ses import (
     AmazonSESInboundWebhookView,
     AmazonSESTrackingWebhookView,
 )
+from .webhooks.brevo import BrevoInboundWebhookView, BrevoTrackingWebhookView
 from .webhooks.mailersend import (
     MailerSendInboundWebhookView,
     MailerSendTrackingWebhookView,
@@ -15,10 +16,6 @@ from .webhooks.postal import PostalInboundWebhookView, PostalTrackingWebhookView
 from .webhooks.postmark import PostmarkInboundWebhookView, PostmarkTrackingWebhookView
 from .webhooks.resend import ResendTrackingWebhookView
 from .webhooks.sendgrid import SendGridInboundWebhookView, SendGridTrackingWebhookView
-from .webhooks.sendinblue import (
-    SendinBlueInboundWebhookView,
-    SendinBlueTrackingWebhookView,
-)
 from .webhooks.sparkpost import (
     SparkPostInboundWebhookView,
     SparkPostTrackingWebhookView,
@@ -31,6 +28,11 @@ urlpatterns = [
         "amazon_ses/inbound/",
         AmazonSESInboundWebhookView.as_view(),
         name="amazon_ses_inbound_webhook",
+    ),
+    path(
+        "brevo/inbound/",
+        BrevoInboundWebhookView.as_view(),
+        name="brevo_inbound_webhook",
     ),
     path(
         "mailersend/inbound/",
@@ -67,11 +69,6 @@ urlpatterns = [
         name="sendgrid_inbound_webhook",
     ),
     path(
-        "sendinblue/inbound/",
-        SendinBlueInboundWebhookView.as_view(),
-        name="sendinblue_inbound_webhook",
-    ),
-    path(
         "sparkpost/inbound/",
         SparkPostInboundWebhookView.as_view(),
         name="sparkpost_inbound_webhook",
@@ -80,6 +77,11 @@ urlpatterns = [
         "amazon_ses/tracking/",
         AmazonSESTrackingWebhookView.as_view(),
         name="amazon_ses_tracking_webhook",
+    ),
+    path(
+        "brevo/tracking/",
+        BrevoTrackingWebhookView.as_view(),
+        name="brevo_tracking_webhook",
     ),
     path(
         "mailersend/tracking/",
@@ -115,11 +117,6 @@ urlpatterns = [
         "sendgrid/tracking/",
         SendGridTrackingWebhookView.as_view(),
         name="sendgrid_tracking_webhook",
-    ),
-    path(
-        "sendinblue/tracking/",
-        SendinBlueTrackingWebhookView.as_view(),
-        name="sendinblue_tracking_webhook",
     ),
     path(
         "sparkpost/tracking/",